cli: simplify writing to stdout
A common method in CLIs is to write to stdout by writing to the lock,
such as `write!(io::stdout(), "foo")`. The benefit of this is that
`write!` returns a `Result`, so if there's an issue with printing, such
as an OOM, then it can be handled or `?`d up to be handled.

The CLI was originally doing this because it was maybe planned to do
something with this, but instead just use `println!`, because this will
never be handled.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>